:bug: fix github pages does not use not-found page bug
